Also run openstackid-release-branch in release
The openstackid-release-branch job seems to produce both versioned
and branch tip artifacts when it's run, so it should be fine to run
this from tag events as well as change merged events. In the future,
the job can be subdivided to only produce one artifact (a versioned
release or an overwritten "latest" branch tip build) depending on
the context.

Change-Id: If9069cbe0b899c3cde8422be2b707df986647a42