Use correct hosts set in post playbooks
Missed in the transition to Zuul v3, there is no longer a legacy
host named "primary" in the Ansible inventory (this was provided for
compatibility with pre-v3 Nodepool multi-node topologies). Switch to
specifying "all" for the applicable hosts in post playbooks, as
there is only one remote node for these jobs anyway so which one
they run on is irrelevant.

Change-Id: I1d66828fd7e9f6c8f4062caa0a794657331e2159